requirements.txt
boto3==1.23.8
botocore==1.26.8
jmespath==1.0.0
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
